requirements.txt
Flask==0.12.2
gunicorn==19.7.1
google-cloud-vision==0.30.1
google-cloud-storage==1.8.0
google-cloud-datastore==1.6.0
google-cloud-language==1.0.2
google-cloud-translate==1.3.1
